requirements.txt
# Proyecto Modular 3 - Simulaciones de Radioterapia con GATE 10 e IA
# Requisitos de Python para el proyecto

# ========================================
# Core Simulation: GATE 10 y OpenGate
# ========================================
opengate>=10.0.0
# opengate_core se instala automáticamente con opengate

# ========================================
# Procesamiento Numérico y Científico
# ========================================
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
pandas>=2.0.0

# ========================================
# Procesamiento de Imágenes Médicas
# ========================================
SimpleITK>=2.2.0
pydicom>=2.3.0
nibabel>=5.1.0

# ========================================
# Deep Learning y IA
# ========================================
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.15.0
# Para modelos avanzados de phase space
# Nota: gaga-phsp 0.9+ no está disponible en PyPI para Python 3.12.
gaga-phsp>=0.7.6  # Generative Adversarial Networks para phase space

# ========================================
# Análisis de Física Médica
# ========================================
pymedphys>=0.39.0  # Incluye herramientas para índice Gamma
# GateTools (si está disponible en PyPI, sino instalar desde source)

# ========================================
# Visualización y Gráficos
# ========================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
vtk>=9.2.0  # Para visualización 3D de geometrías

# ========================================
# Configuración y Gestión de Datos
# ========================================
pyyaml>=6.0
h5py>=3.8.0  # Para almacenamiento eficiente de datos
tqdm>=4.65.0  # Barras de progreso

# ========================================
# Testing y Validación
# ========================================
pytest>=7.3.0
pytest-cov>=4.1.0

# ========================================
# Documentación
# ========================================
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0

# ========================================
# Utilidades
# ========================================
scikit-learn>=1.2.0  # Para preprocesamiento de datos
scikit-image>=0.20.0
pillow>=9.5.0

# ========================================
# Jupyter (opcional, para análisis interactivo)
# ========================================
jupyter>=1.0.0
ipykernel>=6.22.0
notebook>=6.5.0

# ========================================
# Optimización y Paralelización
# ========================================
joblib>=1.2.0
ray>=2.4.0  # Para ejecución distribuida (opcional)

# ========================================
# Notas de Instalación
# ========================================
# 
# Para instalación sin visualización (clusters HPC):
# pip install "opengate[novis]"
#
# Para instalación con CUDA (GPU):
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# Para descargar datos de Geant4:
# export GIT_SSL_NO_VERIFY=1  # Si hay problemas SSL
# opengate_info
#
# Verificar instalación:
# opengate_tests
